Started halted the work on shift option, waiting for a mathematical confirmation

+++ client/src/app/components/app/line/line.component.ts
@@ -48,6 +48,7 @@ export class LineComponent implements OnInit {
 
     this.shortcuts.addShortcut({ keys: 'escape', description: 'Cancel current line' }).subscribe(
       (event) => {
+        if(!this.service.getLineIsDone())
         this.service.deleteLine();
       }
     );
